1st Clean up
